refactor:svg icons
